codecache: Remove cpp_prefix.h duplication per build, then precompile it
Prior to this PR, `_inductor/codegen/cpp_prefix.h` was copied into a new temporary directory on every inductor run utilizing the CPP backend, then included in the output source code. Instead, this puts it in an appropriate place in the torch includes, and includes it from there. This allows us to precompile it in cpp_wrapper and AOT inductor mode, saving significant compilation time.

Note for reviewers: I believe I've traced down all the implications of this change for FBCode here on the OSS side, but someone will need to run this through tests on the FB side to be sure.

ghstack-source-id: 1c13ea87718016c38fce6cccf29fb0d5b812ee32
Pull Request resolved: https://github.com/pytorch/pytorch/pull/144293